Indexing functions for subsignal-extraction

Indexing functions have the following signature: function(x, ...). x denotes a signal-values vector (see tbl_mbte for details). In other words, the measured values of a signal get passed as a numeric vector. The job of an indexing function is to find the start- and end-indices of a subsignal (used by mbte_extract_subsignals).

Value

A list with the following names should be returned:

start

Start indices of the found subsignals.

end

End indices of the found subsignals.

Note

The length of the returned indices must match the number of found subsignals (e.g. 3 Subsignals found ==> length of start- and end-indices vectors must equal 3). The indices-vectors have to be of type integer.