Perform unnesting of fitting data

The `fits`-list column gets unnested. This function is useful if the fitted signal-values are processed further.

mbte_unnest_fits(x)

Arguments

x
A tbl_mbte.

Value

A tbl_mbte with the following columns:

desc

`desc` is a placeholder for all columns except the `signal`- and the `fits`-columns. Those columns can be seen as "descriptive" columns describing the measured signals.

fit

The name of the method used for fitting (character). See mbte_fit for details.

time

The time-column unnested from the `signal`-column.

value

The values of the predicted signals (originally in `signal`- column).

See also

Examples

# load sample dataset
data(filtered_signals)
filtered_signals

#> # A tibble: 37 x 3
#>    mv    signal_nr signal           
#>    <chr>     <int> <list>           
#>  1 mv1           1 <tibble [22 × 2]>
#>  2 mv3           2 <tibble [22 × 2]>
#>  3 mv5           1 <tibble [30 × 2]>
#>  4 mv6           1 <tibble [21 × 2]>
#>  5 mv6           3 <tibble [30 × 2]>
#>  6 mv7           1 <tibble [56 × 2]>
#>  7 mv8           1 <tibble [29 × 2]>
#>  8 mv8           2 <tibble [26 × 2]>
#>  9 mv9           1 <tibble [63 × 2]>
#> 10 mv10          1 <tibble [25 × 2]>
#> # ... with 27 more rows

# perform fitting
fitted <- mbte_fit(filtered_signals, lm = lm(value ~ t, .signal))

# unnest predictions for signal-values (==> fitted signals)
#
# NOTE: "lm" is the only method used for fitting and therefore the `fit`-
# column consists only of "lm"
mbte_unnest_fits(fitted)

#> # A tibble: 1,367 x 5
#>    mv    signal_nr fit       t value
#>    <chr>     <int> <chr> <int> <dbl>
#>  1 mv1           1 lm       29 0.302
#>  2 mv1           1 lm       30 0.344
#>  3 mv1           1 lm       31 0.386
#>  4 mv1           1 lm       32 0.429
#>  5 mv1           1 lm       33 0.471
#>  6 mv1           1 lm       34 0.514
#>  7 mv1           1 lm       35 0.556
#>  8 mv1           1 lm       36 0.599
#>  9 mv1           1 lm       37 0.641
#> 10 mv1           1 lm       38 0.684
#> # ... with 1,357 more rows