The main data-structure of the mbte-package

A tbl_mbte is a tibble under the hood. Additionally, attributes are stored, which contain column-names. All mentioned column names are stored as symbols.

Columns

The following columns are used during the analysis (the attributes storing the column names have the same names).

time

A numeric column containing the time of a specific measurement.

value

A numeric column containing the measured value of a signal (e.g. intensity).

signal

The time- and value-column combined into a list-column (e.g. by using mbte_nest_signals). Every element of this column is a tibble containing the time- and value-column of the signal.

fits

This list-column also contains tibbles. The fitted signal-values are stored as columns. NOTE: Every contained tibble has the same number of rows as the signal-tibble of the same row. This column may be generated using mbte_fit.

See also